requirements.txt
asgiref==3.2.10
Django==3.0.8
django-stdimage==5.1.1
Pillow==7.2.0
progressbar2==3.51.4
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
